+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Support for port range syntax (e.g. `8080-8085`) in `proxy.port` and `proxy.ports` consul tags
 
 ### Changed